+++ .github/CODEOWNERS
@@ -256,7 +256,7 @@ models/demos/utils/llm_demo_utils.py @skhorasganiTT @mtairum @uaydonat
 models/experimental/stable_diffusion_xl_base @mbezuljTT @pavlepopovic @ipotkonjak-tt @tenstorrent/metalium-developers-convolutions
 models/demos/mobilenetv2 @tenstorrent/cse-developer-ttnn
 models/demos/yolov8s_world @tenstorrent/cse-developer-ttnn
-models/experimental/yolov10x @tenstorrent/cse-developer-ttnn
+models/demos/yolov10x @tenstorrent/cse-developer-ttnn
 models/demos/yolov9c @tenstorrent/cse-developer-ttnn
 models/demos/yolov8x @tenstorrent/cse-developer-ttnn
 models/demos/yolov8s @tenstorrent/cse-developer-ttnn
