+++ models/demos/yolov10x/demo/images/dog.jpg
