+++ models/demos/yolov10x/web_demo/client/run_on_client_YOLOv10_Metal
@@ -1,0 +1,2 @@
+#!/bin/bash
+streamlit run yolov10.py -- "$@"
