+++ models/demos/yolov10x/web_demo/server/requirements.txt
@@ -1,0 +1,5 @@
+fastapi==0.85.1
+uvicorn==0.19.0
+python-multipart==0.0.5
+
+-f https://download.pytorch.org/whl/cpu/torch_stable.html
