--- models/experimental/yolov10x/reference/yolov10x.py
@@ -1,367 +1,0 @@
-# SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
-
-# SPDX-License-Identifier: Apache-2.0
-
-import torch
-import torch.nn as nn
-import copy
-
-
-def dist2bbox(distance, anchor_points, xywh=True, dim=-1):
-    lt, rb = distance.chunk(2, dim)
-    x1y1 = anchor_points - lt
-    x2y2 = anchor_points + rb
-    if xywh:
-        c_xy = (x1y1 + x2y2) / 2
-        wh = x2y2 - x1y1
-        return torch.cat((c_xy, wh), dim)
-    return torch.cat((x1y1, x2y2), dim)
-
-
-def make_anchors(feats, strides, grid_cell_offset=0.5):
-    anchor_points, stride_tensor = [], []
-    assert feats is not None
-    dtype, device = feats[0].dtype, feats[0].device
-    for i, stride in enumerate(strides):
-        h, w = feats[i].shape[2:] if isinstance(feats, list) else (int(feats[i][0]), int(feats[i][1]))
-        sx = torch.arange(end=w, device=device, dtype=dtype) + grid_cell_offset
-        sy = torch.arange(end=h, device=device, dtype=dtype) + grid_cell_offset
-        sy, sx = torch.meshgrid(sy, sx, indexing="ij")
-        anchor_points.append(torch.stack((sx, sy), -1).view(-1, 2))
-        stride_tensor.append(torch.full((h * w, 1), stride, dtype=dtype, device=device))
-    return torch.cat(anchor_points), torch.cat(stride_tensor)
-
-
-def autopad(k, p=None, d=1):
-    if d > 1:
-        k = d * (k - 1) + 1 if isinstance(k, int) else [d * (x - 1) + 1 for x in k]
-    if p is None:
-        p = k // 2 if isinstance(k, int) else [x // 2 for x in k]
-    return p
-
-
-class Concat(nn.Module):
-    def __init__(self, dimension=1):
-        super().__init__()
-        self.d = dimension
-
-    def forward(self, x):
-        out = torch.cat(x, self.d)
-        return out
-
-
-class Conv(nn.Module):
-    default_act = nn.SiLU()
-
-    def __init__(self, c1, c2, k=1, s=1, p=None, g=1, d=1, act=True):
-        super().__init__()
-        self.conv = nn.Conv2d(c1, c2, k, s, autopad(k, p, d), groups=g, dilation=d, bias=False)
-        self.bn = nn.BatchNorm2d(c2, eps=0.001, momentum=0.03)
-        self.act = self.default_act if act is True else act if isinstance(act, nn.Module) else nn.Identity()
-
-    def forward(self, x):
-        return self.act(self.bn(self.conv(x)))
-
-
-class Bottleneck(nn.Module):
-    def __init__(self, c1, c2, shortcut=True, g=1, k=(3, 3), e=0.5):
-        super().__init__()
-        c_ = int(c2 * e)
-        self.cv1 = Conv(c1, c_, k[0], 1)
-        self.cv2 = Conv(c_, c2, k[1], 1, g=g)
-        self.add = shortcut and c1 == c2
-
-    def forward(self, x):
-        return x + self.cv2(self.cv1(x)) if self.add else self.cv2(self.cv1(x))
-
-
-class C2f(nn.Module):
-    def __init__(self, c1, c2, n=1, shortcut=False, g=1, e=0.5):
-        super().__init__()
-        self.c = int(c2 * e)
-        self.cv1 = Conv(c1, 2 * self.c, 1, 1)
-        self.cv2 = Conv((2 + n) * self.c, c2, 1)
-        self.m = nn.ModuleList(Bottleneck(self.c, self.c, shortcut, g, k=((3, 3), (3, 3)), e=1.0) for _ in range(n))
-
-    def forward(self, x):
-        y = list(self.cv1(x).chunk(2, 1))
-        y.extend(m(y[-1]) for m in self.m)
-        return self.cv2(torch.cat(y, 1))
-
-    def forward_split(self, x):
-        y = self.cv1(x).split((self.c, self.c), 1)
-        y = [y[0], y[1]]
-        y.extend(m(y[-1]) for m in self.m)
-        return self.cv2(torch.cat(y, 1))
-
-
-class SPPF(nn.Module):
-    def __init__(self, c1, c2, k=5):
-        super().__init__()
-        c_ = c1 // 2
-        self.cv1 = Conv(c1, c_, 1, 1)
-        self.cv2 = Conv(c_ * 4, c2, 1, 1)
-        self.m = nn.MaxPool2d(kernel_size=k, stride=1, padding=k // 2)
-
-    def forward(self, x):
-        y = [self.cv1(x)]
-        y.extend(self.m(y[-1]) for _ in range(3))
-        return self.cv2(torch.cat(y, 1))
-
-
-class CIB(nn.Module):
-    def __init__(self, c1, c2, shortcut=True, e=0.5, lk=False):
-        super().__init__()
-        c_ = int(c2 * e)
-        self.cv1 = nn.Sequential(
-            Conv(c1, c1, 3, g=c1),
-            Conv(c1, 2 * c_, 1),
-            RepVGGDW(2 * c_) if lk else Conv(2 * c_, 2 * c_, 3, g=2 * c_),
-            Conv(2 * c_, c2, 1),
-            Conv(c2, c2, 3, g=c2),
-        )
-
-        self.add = shortcut and c1 == c2
-
-    def forward(self, x):
-        return x + self.cv1(x) if self.add else self.cv1(x)
-
-
-class SCDown(nn.Module):
-    def __init__(self, c1, c2, k, s):
-        super().__init__()
-        self.cv1 = Conv(c1, c2, 1, 1)
-        self.cv2 = Conv(c2, c2, k=k, s=s, g=c2, act=False)
-
-    def forward(self, x):
-        return self.cv2(self.cv1(x))
-
-
-class C2fCIB(C2f):
-    def __init__(self, c1, c2, n=1, shortcut=False, lk=False, g=1, e=0.5):
-        super().__init__(c1, c2, n, shortcut, g, e)
-        self.m = nn.ModuleList(CIB(self.c, self.c, shortcut, e=1.0, lk=lk) for _ in range(n))
-
-
-class Attention(nn.Module):
-    def __init__(self, dim, num_heads=8, attn_ratio=0.5):
-        super().__init__()
-        self.num_heads = num_heads
-        self.head_dim = dim // num_heads
-        self.key_dim = int(self.head_dim * attn_ratio)
-        self.scale = self.key_dim**-0.5
-        nh_kd = self.key_dim * num_heads
-        h = dim + nh_kd * 2
-        self.qkv = Conv(dim, h, 1, act=False)
-        self.proj = Conv(dim, dim, 1, act=False)
-        self.pe = Conv(dim, dim, 3, 1, g=dim, act=False)
-
-    def forward(self, x):
-        B, C, H, W = x.shape
-        N = H * W
-        qkv = self.qkv(x)
-        q, k, v = qkv.view(B, self.num_heads, self.key_dim * 2 + self.head_dim, N).split(
-            [self.key_dim, self.key_dim, self.head_dim], dim=2
-        )
-
-        attn = (q.transpose(-2, -1) @ k) * self.scale
-        attn = attn.softmax(dim=-1)
-        x = (v @ attn.transpose(-2, -1)).view(B, C, H, W) + self.pe(v.reshape(B, C, H, W))
-        x = self.proj(x)
-        return x
-
-
-class PSA(nn.Module):
-    def __init__(self, c1, c2, e=0.5):
-        super().__init__()
-        assert c1 == c2
-        self.c = int(c1 * e)
-        self.cv1 = Conv(c1, 2 * self.c, 1, 1)
-        self.cv2 = Conv(2 * self.c, c1, 1)
-
-        self.attn = Attention(self.c, attn_ratio=0.5, num_heads=self.c // 64)
-        self.ffn = nn.Sequential(Conv(self.c, self.c * 2, 1), Conv(self.c * 2, self.c, 1, act=False))
-
-    def forward(self, x):
-        a, b = self.cv1(x).split((self.c, self.c), dim=1)
-        b = b + self.attn(b)
-        b = b + self.ffn(b)
-        return self.cv2(torch.cat((a, b), 1))
-
-
-class DFL(nn.Module):
-    def __init__(self, c1=16):
-        super().__init__()
-        self.conv = nn.Conv2d(c1, 1, 1, bias=False).requires_grad_(False)
-        x = torch.arange(c1, dtype=torch.float)
-        self.conv.weight.data[:] = nn.Parameter(x.view(1, c1, 1, 1))
-        self.c1 = c1
-
-    def forward(self, x):
-        b, _, a = x.shape
-        return self.conv(x.view(b, 4, self.c1, a).transpose(2, 1).softmax(1)).view(b, 4, a)
-
-
-class Detect(nn.Module):
-    dynamic = False
-    export = False
-    format = None
-    end2end = False
-    max_det = 300
-    shape = None
-    anchors = torch.empty(0)
-    strides = torch.empty(0)
-    legacy = False
-
-    def __init__(self, nc=80, ch=()):
-        super().__init__()
-        self.nc = nc
-        self.nl = len(ch)
-        self.reg_max = 16
-        self.no = nc + self.reg_max * 4
-        self.stride = torch.tensor([8.0, 16.0, 32.0])
-        c2, c3 = max((16, ch[0] // 4, self.reg_max * 4)), max(ch[0], min(self.nc, 100))
-        self.cv2 = nn.ModuleList(
-            nn.Sequential(Conv(x, c2, 3), Conv(c2, c2, 3), nn.Conv2d(c2, 4 * self.reg_max, 1)) for x in ch
-        )
-        self.cv3 = nn.ModuleList(nn.Sequential(Conv(x, c3, 3), Conv(c3, c3, 3), nn.Conv2d(c3, self.nc, 1)) for x in ch)
-        self.dfl = DFL(self.reg_max) if self.reg_max > 1 else nn.Identity()
-
-        if self.end2end:
-            self.one2one_cv2 = copy.deepcopy(self.cv2)
-            self.one2one_cv3 = copy.deepcopy(self.cv3)
-
-    def forward(self, x):
-        if self.end2end:
-            return self.forward_end2end(x)
-
-    def forward_end2end(self, x):
-        x_detach = [xi.detach() for xi in x]
-        one2one = [
-            torch.cat((self.one2one_cv2[i](x_detach[i]), self.one2one_cv3[i](x_detach[i])), 1) for i in range(self.nl)
-        ]
-        y = self._inference(one2one)
-        return y
-
-    def _inference(self, x):
-        shape = x[0].shape
-        x_cat = torch.cat([xi.view(shape[0], self.no, -1) for xi in x], 2)
-        if self.format != "imx" and (self.dynamic or self.shape != shape):
-            self.anchors, self.strides = (x.transpose(0, 1) for x in make_anchors(x, self.stride, 0.5))
-            self.shape = shape
-
-        box, cls = x_cat.split((self.reg_max * 4, self.nc), 1)
-
-        dbox = self.decode_bboxes(self.dfl(box), self.anchors.unsqueeze(0)) * self.strides
-
-        return torch.cat((dbox, cls.sigmoid()), 1)
-
-    def decode_bboxes(self, bboxes, anchors, xywh=True):
-        return dist2bbox(bboxes, anchors, xywh=xywh and (not self.end2end), dim=1)
-
-
-class v10Detect(Detect):
-    end2end = True
-
-    def __init__(self, nc=80, ch=()):
-        super().__init__(nc, ch)
-        c3 = max(ch[0], min(self.nc, 100))
-        self.cv3 = nn.ModuleList(
-            nn.Sequential(
-                nn.Sequential(Conv(x, x, 3, g=x), Conv(x, c3, 1)),
-                nn.Sequential(Conv(c3, c3, 3, g=c3), Conv(c3, c3, 1)),
-                nn.Conv2d(c3, self.nc, 1),
-            )
-            for x in ch
-        )
-        self.one2one_cv3 = copy.deepcopy(self.cv3)
-
-
-class YOLOv10(nn.Module):
-    def __init__(self, nc=80, anchors=()):
-        super().__init__()
-        self.model = nn.Sequential(
-            Conv(3, 80, 3, 2, 1),
-            Conv(80, 160, 3, 2, 1),
-            C2f(160, 160, n=3, shortcut=True),
-            Conv(160, 320, 3, 2, 1),
-            C2f(320, 320, n=6, shortcut=True),
-            SCDown(320, 640, 3, 2),
-            C2fCIB(640, 640, n=6, shortcut=True),
-            SCDown(640, 640, 3, 2),
-            C2fCIB(640, 640, n=3, shortcut=True),
-            SPPF(640, 640),
-            PSA(640, 640),
-            nn.Upsample(None, 2, "nearest"),
-            Concat(),
-            C2fCIB(1280, 640, n=3, shortcut=True),
-            nn.Upsample(None, 2, "nearest"),
-            Concat(),
-            C2f(960, 320, n=3),
-            Conv(320, 320, 3, 2, 1),
-            Concat(),
-            C2fCIB(960, 640, n=3, shortcut=True),
-            SCDown(640, 640, 3, 2),
-            Concat(),
-            C2fCIB(1280, 640, n=3, shortcut=True),
-            v10Detect(nc=80, ch=(320, 640, 640)),
-        )
-
-    def forward(self, x):
-        x = self.model[0](x)
-        x = self.model[1](x)
-        x = self.model[2](x)
-        x = self.model[3](x)
-        x = self.model[4](x)
-        x4 = x
-        x = self.model[5](x)
-        x = self.model[6](x)
-        x6 = x
-        x = self.model[7](x)
-        x = self.model[8](x)
-        x = self.model[9](x)
-        x = self.model[10](x)
-        x10 = x
-        x = self.model[11](x)
-        x = torch.cat((x, x6), 1)
-        x = self.model[13](x)
-        x13 = x
-        x = self.model[14](x)
-        x = torch.cat([x, x4], 1)
-        x = self.model[16](x)
-        x16 = x
-        x = self.model[17](x)
-        x = torch.cat([x, x13], 1)
-        x = self.model[19](x)
-        x19 = x
-        x = self.model[20](x)
-        x = torch.cat([x, x10], 1)
-        x = self.model[22](x)
-        x22 = x
-        x = self.model[23]([x16, x19, x22])
-
-        return x
-
-
-class BaseModel(nn.Module):
-    def forward(self, x, *args, **kwargs):
-        if isinstance(x, dict):
-            return self.loss(x, *args, **kwargs)
-        return self.predict(x, *args, **kwargs)
-
-    def predict(self, x, profile=False, visualize=False, augment=False, embed=None):
-        return self._predict_once(x, profile, visualize, embed)
-
-    def _predict_once(self, x, profile=False, visualize=False, embed=None):
-        y, dt, embeddings = [], [], []
-        for m in self.model:
-            if m.f != -1:
-                x = y[m.f] if isinstance(m.f, int) else [x if j == -1 else y[j] for j in m.f]
-            x = m(x)
-            y.append(x if m.i in self.save else None)
-        return x
-
-
-class DetectionModel(BaseModel):
-    def __init__(self, cfg="yolov10x.yaml", ch=3, nc=None, verbose=True):
-        super().__init__()
