--- models/experimental/yolov10x/runner/performant_runner.py
@@ -1,118 +1,0 @@
-# SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
-
-# SPDX-License-Identifier: Apache-2.0
-
-
-import ttnn
-from models.experimental.yolov10x.runner.performant_runner_infra import YOLOv10PerformanceRunnerInfra
-from tests.ttnn.utils_for_testing import assert_with_pcc
-
-
-class YOLOv10PerformantRunner:
-    def __init__(
-        self,
-        device,
-        device_batch_size=1,
-        act_dtype=ttnn.bfloat16,
-        weight_dtype=ttnn.bfloat16,
-        model_location_generator=None,
-        resolution=(640, 640),
-        torch_input_tensor=None,
-        use_pretrained_weight=True,
-    ):
-        self.device = device
-        self.resolution = resolution
-        self.torch_input_tensor = torch_input_tensor
-        self.runner_infra = YOLOv10PerformanceRunnerInfra(
-            device,
-            device_batch_size,
-            act_dtype,
-            weight_dtype,
-            model_location_generator,
-            resolution=resolution,
-            torch_input_tensor=self.torch_input_tensor,
-            use_pretrained_weight=use_pretrained_weight,
-        )
-
-        (
-            self.tt_inputs_host,
-            sharded_mem_config_DRAM,
-            self.input_mem_config,
-        ) = self.runner_infra.setup_dram_sharded_input(device)
-        self.tt_image_res = self.tt_inputs_host.to(device, sharded_mem_config_DRAM)
-
-        self._capture_yolov10_trace_2cqs()
-
-    def _capture_yolov10_trace_2cqs(self):
-        # Initialize the op event so we can write
-        self.op_event = ttnn.record_event(self.device, 0)
-
-        # First run configures convs JIT
-        ttnn.wait_for_event(1, self.op_event)
-        ttnn.copy_host_to_device_tensor(self.tt_inputs_host, self.tt_image_res, 1)
-        self.write_event = ttnn.record_event(self.device, 1)
-        ttnn.wait_for_event(0, self.write_event)
-        self.runner_infra.input_tensor = ttnn.to_memory_config(self.tt_image_res, self.input_mem_config)
-        spec = self.runner_infra.input_tensor.spec
-        self.op_event = ttnn.record_event(self.device, 0)
-        self.runner_infra.run()
-        self.runner_infra.validate()
-        self.runner_infra.dealloc_output()
-
-        # Optimized run
-        ttnn.wait_for_event(1, self.op_event)
-        ttnn.copy_host_to_device_tensor(self.tt_inputs_host, self.tt_image_res, 1)
-        self.write_event = ttnn.record_event(self.device, 1)
-        ttnn.wait_for_event(0, self.write_event)
-        self.runner_infra.input_tensor = ttnn.to_memory_config(self.tt_image_res, self.input_mem_config)
-        self.op_event = ttnn.record_event(self.device, 0)
-        self.runner_infra.run()
-        self.runner_infra.validate()
-
-        # Capture
-        ttnn.wait_for_event(1, self.op_event)
-        ttnn.copy_host_to_device_tensor(self.tt_inputs_host, self.tt_image_res, 1)
-        self.write_event = ttnn.record_event(self.device, 1)
-        ttnn.wait_for_event(0, self.write_event)
-        self.runner_infra.input_tensor = ttnn.to_memory_config(self.tt_image_res, self.input_mem_config)
-        self.op_event = ttnn.record_event(self.device, 0)
-        self.runner_infra.dealloc_output()
-        trace_input_addr = self.runner_infra.input_tensor.buffer_address()
-        self.tid = ttnn.begin_trace_capture(self.device, cq_id=0)
-        self.runner_infra.run()
-        self.input_tensor = ttnn.allocate_tensor_on_device(spec, self.device)
-        ttnn.end_trace_capture(self.device, self.tid, cq_id=0)
-        assert trace_input_addr == self.input_tensor.buffer_address()
-
-    def _execute_yolov10_trace_2cqs_inference(self, tt_inputs_host=None):
-        tt_inputs_host = self.tt_inputs_host if tt_inputs_host is None else tt_inputs_host
-        ttnn.wait_for_event(1, self.op_event)
-        ttnn.copy_host_to_device_tensor(tt_inputs_host, self.tt_image_res, 1)
-        self.write_event = ttnn.record_event(self.device, 1)
-        ttnn.wait_for_event(0, self.write_event)
-        # TODO: Add in place support to ttnn to_memory_config
-        self.input_tensor = ttnn.reshard(self.tt_image_res, self.input_mem_config, self.input_tensor)
-        self.op_event = ttnn.record_event(self.device, 0)
-        ttnn.execute_trace(self.device, self.tid, cq_id=0, blocking=False)
-        ttnn.synchronize_device(self.device)
-
-        ttnn_output_tensor = self.runner_infra.output_tensor
-        return ttnn_output_tensor  # get_model_result(ttnn_output_tensor, self.resolution)
-
-    def _validate(self, input_tensor, result_output_tensor):
-        torch_output_tensor = self.runner_infra.torch_output_tensor
-        assert_with_pcc(torch_output_tensor, result_output_tensor, 0.99)
-
-    def run(self, torch_input_tensor, check_pcc=False):
-        n, c, h, w = torch_input_tensor.shape
-        tt_inputs_host, _ = self.runner_infra._setup_l1_sharded_input(self.device, torch_input_tensor)
-
-        output = self._execute_yolov10_trace_2cqs_inference(tt_inputs_host)
-
-        if check_pcc:
-            self._validate(torch_input_tensor, output)
-
-        return output
-
-    def release(self):
-        ttnn.release_trace(self.device, self.tid)
